requirements.txt
# ViralOps Engine — Dependencies
# EMADS-PR v1.0

# ── LangGraph / LLM ──
langgraph>=0.2.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
openai>=1.50.0

# ── Config ──
pyyaml>=6.0
python-dotenv>=1.0

# ── HTTP / Network ──
httpx>=0.27.0
aiohttp>=3.9.0

# ── Data ──
pydantic>=2.0
structlog>=24.0

# ── Web Dashboard ──
fastapi>=0.115.0
uvicorn>=0.30.0
jinja2>=3.1
python-multipart>=0.0.9

# ── RSS Feed ──
feedparser>=6.0

# ── Token counting ──
tiktoken>=0.7.0

# ── Testing ──
pytest>=8.0
pytest-asyncio>=0.23
